Improve conversation captcha failure message
The existing conversation capthca failure message is very specific to the default captcha 'Incorrect image validation code. Please check the image and re-enter the letters or numbers as necessary.'

This change improves it by making it more generic, so it is relevant to any captcha.

There is scope for better capthca error responses here. This is just a minimal fix.